+++ packages/core/src/components/StoreUpdater/index.tsx
@@ -21,6 +21,7 @@ type StoreUpdaterProps = Pick<
   | 'nodesConnectable'
   | 'nodesFocusable'
   | 'edgesFocusable'
+  | 'edgesUpdatable'
   | 'minZoom'
   | 'maxZoom'
   | 'nodeExtent'
@@ -99,6 +100,7 @@ const StoreUpdater = ({
   nodesConnectable,
   nodesFocusable,
   edgesFocusable,
+  edgesUpdatable,
   elevateNodesOnSelect,
   minZoom,
   maxZoom,
@@ -163,6 +165,7 @@ const StoreUpdater = ({
   useDirectStoreUpdater('nodesConnectable', nodesConnectable, store.setState);
   useDirectStoreUpdater('nodesFocusable', nodesFocusable, store.setState);
   useDirectStoreUpdater('edgesFocusable', edgesFocusable, store.setState);
+  useDirectStoreUpdater('edgesUpdatable', edgesUpdatable, store.setState);
   useDirectStoreUpdater('elementsSelectable', elementsSelectable, store.setState);
   useDirectStoreUpdater('elevateNodesOnSelect', elevateNodesOnSelect, store.setState);
   useDirectStoreUpdater('snapToGrid', snapToGrid, store.setState);