+++ packages/core/src/container/ZoomPane/index.tsx
@@ -139,7 +139,7 @@ const ZoomPane = ({
             -(deltaX / currentZoom) * panOnScrollSpeed,
             -(deltaY / currentZoom) * panOnScrollSpeed
           );
-        });
+        }, { passive: false });
       } else if (typeof d3ZoomHandler !== 'undefined') {
         d3Selection.on('wheel.zoom', function (event: any, d: any) {
           if (!preventScrolling || isWrappedWithClass(event, noWheelClassName)) {
@@ -148,7 +148,7 @@ const ZoomPane = ({
 
           event.preventDefault();
           d3ZoomHandler.call(this, event, d);
-        });
+        }, { passive: false });
       }
     }
   }, [